El historial de pedidos ya ordena por fecha y total

+++ src/model/OrderValueCompare.java
@@ -6,7 +6,7 @@ public class OrderValueCompare implements Comparator<Order> {
 
 
 	public int compare(Order o1, Order o2) {
-		return o1.getTotalPrice() - o2.getTotalPrice();
+		return (int) o1.getTotalPrice() - o2.getTotalPrice();
 	}
